requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
pyqtgraph>=0.13.3
pywin32>=306
Pillow>=10.2.0
